+++ examples/example.c
@@ -579,8 +579,8 @@ main(int argc, char **argv)
         sentry_options_set_cache_max_age(options, 5 * 24 * 60 * 60); // 5 days
         sentry_options_set_cache_max_items(options, 5);
     }
-    if (has_arg(argc, argv, "http-retry")) {
-        sentry_options_set_http_retries(options, 5);
+    if (has_arg(argc, argv, "no-http-retry")) {
+        sentry_options_set_http_retry(options, false);
     }
 
     if (has_arg(argc, argv, "enable-metrics")) {