+++ src/sentry_options.c
@@ -75,6 +75,7 @@ sentry_options_new(void)
     opts->traces_sample_rate = 0.0;
     opts->max_spans = SENTRY_SPANS_MAX;
     opts->handler_strategy = SENTRY_HANDLER_STRATEGY_DEFAULT;
+    opts->http_retry = true;
 
     return opts;
 }
@@ -829,15 +830,15 @@ sentry_options_set_handler_strategy(
 #endif // SENTRY_PLATFORM_LINUX
 
 void
-sentry_options_set_http_retries(sentry_options_t *opts, int http_retries)
+sentry_options_set_http_retry(sentry_options_t *opts, int enabled)
 {
-    opts->http_retries = http_retries;
+    opts->http_retry = enabled;
 }
 
 int
-sentry_options_get_http_retries(const sentry_options_t *opts)
+sentry_options_get_http_retry(const sentry_options_t *opts)
 {
-    return opts->http_retries;
+    return opts->http_retry;
 }
 
 void